requirements.txt
pymongo~=4.8.0
Telethon~=1.36.0
mysql-connector-python~=9.0.0
pika~=1.3.2
aiogram~=3.12.0
aiohttp~=3.10.3
vedis~=0.7.2